moves usage samples to examples folder
