+++ examples/config.js
